+++ api/src/main/java/io/jsonwebtoken/CompressionCodecResolver.java
@@ -41,6 +41,6 @@ public interface CompressionCodecResolver {
      * @return CompressionCodec matching the {@code zip} header, or null if there is no {@code zip} header.
      * @throws CompressionException if a {@code zip} header value is found and not supported.
      */
-    CompressionCodec resolveCompressionCodec(Header header) throws CompressionException;
+    CompressionCodec resolveCompressionCodec(Header<?> header) throws CompressionException;
 
 }