+++ impl/src/main/java/io/jsonwebtoken/impl/CompressionCodecLocator.java
@@ -3,10 +3,11 @@ package io.jsonwebtoken.impl;
 import io.jsonwebtoken.CompressionCodec;
 import io.jsonwebtoken.CompressionCodecResolver;
 import io.jsonwebtoken.Header;
+import io.jsonwebtoken.Locator;
 import io.jsonwebtoken.impl.lang.Function;
 import io.jsonwebtoken.lang.Assert;
 
-public class CompressionCodecLocator implements Function<Header<?>, CompressionCodec> {
+public class CompressionCodecLocator implements Function<Header<?>, CompressionCodec>, Locator<CompressionCodec> {
 
     private final CompressionCodecResolver resolver;
 
@@ -17,5 +18,10 @@ public class CompressionCodecLocator implements Function<Header<?>, CompressionC
     @Override
     public CompressionCodec apply(Header<?> header) {
         return resolver.resolveCompressionCodec(header);
+    }
+
+    @Override
+    public CompressionCodec locate(Header<?> header) {
+        return apply(header);
     }
 }