+++ impl/src/test/groovy/io/jsonwebtoken/impl/compression/DefaultCompressionCodecResolverTest.groovy
@@ -19,6 +19,7 @@ import io.jsonwebtoken.CompressionCodec
 import io.jsonwebtoken.CompressionCodecs
 import io.jsonwebtoken.CompressionException
 import io.jsonwebtoken.impl.DefaultHeader
+import io.jsonwebtoken.impl.DefaultJwsHeader
 import io.jsonwebtoken.impl.io.FakeServiceDescriptorClassLoader
 import io.jsonwebtoken.impl.lang.Services
 import org.junit.Assert
@@ -51,25 +52,15 @@ class DefaultCompressionCodecResolverTest {
     }
 
     @Test
-    void overrideDefaultCompressionImplTest() {
+    void testCustomCompressionCodecServiceDoesNotOverrideStandardCodecs() {
         FakeServiceDescriptorClassLoader.runWithFake "io.jsonwebtoken.io.compression.CompressionCodec.test.override", {
 
             // first make sure the service loader actually resolves the test class
             assertThat Services.loadAll(CompressionCodec), hasItem(instanceOf(YagCompressionCodec))
 
+            def header = new DefaultJwsHeader(['zip': 'gzip'])
             // now we know the class is loadable, make sure we ALWAYS return the GZIP impl
-            assertThat new DefaultCompressionCodecResolver().byName("gzip"), instanceOf(GzipCompressionCodec)
-        }
-    }
-
-    @Test
-    void emptyCompressionAlgInHeaderTest() {
-        //noinspection GroovyUnusedCatchParameter
-        try {
-            new DefaultCompressionCodecResolver().byName("")
-            Assert.fail("Expected IllegalArgumentException to be thrown")
-        } catch (IllegalArgumentException e) {
-            // expected
+            assertThat new DefaultCompressionCodecResolver().locate(header), instanceOf(GzipCompressionCodec)
         }
     }
 }